app/AuthAPI/requirements.txt
fastapi
uvicorn
cassandra-driver
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
email-validator
